`safe_free` function added
Added the function `safe_free_function` and the macro `safe_free`. The function `safe_free_function` requires the user the cast the address of the pointer to be freed to (void **). The macro does this for the user, so the user only needs to call `safe_free(pointer)`.
Also added a define for `safe_malloc` that includes the function name of the calling function.