`safe_realloc` function added
Added the function `safe_realloc_function` and the macro `safe_realloc`.
The purpose is to provide memory reallocation with strict boundary and
error checking. The function `safe_realloc_function` requires the user to
pass the address of the pointer to be reallocated and cast it to (void**),
it also needs to know the name of the calling function. The macro
`safe_realloc` performs these additional steps, so the user can call
`safe_realloc` the same way as calling `realloc`.