Download Criterion release instead of building
Instead of keeping Criterion as a submodule and build it every time on
Travis and Appveyor, download the latest release (v2.2.2 during time of
writing) and use that. This reduces download time, build time and build
log length.
The reason for using a submodule before, was that Criterion version v2.2.1
generated warnings when compiling the tests, along with other issues that
are fixed in v2.2.2.