`safe_calloc` function added
Added the function `safe_calloc_function` and the macro `safe_calloc`.
The purpose is to provide memory reallocation with strict boundary and
error checking. The function `safe_calloc_function` requires the user to
pass the name of the calling function, so this can be displayed if
something goes wrong. The macro `safe_calloc` performs this additional
step, so the user can call `safe_calloc` the same way as calling `calloc`.

Unit test for this function are included with this commit.