Add fallback `safe_free` to test in safe_realloc
Passing NULL to the function `safe_realloc_function` will cause it to fail
and call exit. But if the code should somehow change and this behaviour
would be broken, then the there will be a memory leak in the test case
`address_null_size_valid_expect_exit` in `safe_realloc`. Adding a
`safe_free` call fixes this problem, should it occur.