Enable more warnings on the Windows build
For MSVC we now use the /Wall option instead of /W4. This option enables
various usefull warnings, but also notification-like warnings that cause
the build to fail when using /WX, so we disable those. A few of these
warnings are triggered by Criterion when building the tests, so these are
only disabled when building the tests.
Also improved CMakeLists.txt layout.