Replace `safe_malloc` with `el_malloc` functions
The `el_malloc` functions are more flexible in their usage.
`el_unsafe_malloc` is the root of the new functions. It is marked is
unsafe because direct usage is not recommended. The wrapper macros are
better suited for that.
It is possible to use `el_malloc` as a malloc replacement, the difference
is that `el_malloc` exits on errors and prints the file and line from
which is was called. Error reporting and exit behaviour can be customized
when using `el_custom_malloc`.
The `el_malloc` functions also initialize the newly allocated memory to
all zeros.

Also passing the -std=c99 flags to GCC.