Replace `safe_calloc` with `el_calloc` functions
The `el_calloc` functions are more flexible in their usage.
`el_unsafe_calloc` is the root of the new functions. It is marked is
unsafe because direct usage is not recommended. The wrapper macros are
better suited for that.
It is possible to use `el_calloc` as a calloc replacement, the difference
is that `el_calloc` exits on errors and prints the file and line from
which is was called. Error reporting and exit behaviour can be customized
when using `el_custom_calloc`.
The `el_calloc` functions also initialize the newly allocated memory to
all zeros, just like regular `calloc.`