Replace `safe_realloc` with `el_realloc` functions
The `el_realloc` functions are more flexible in their usage.
`el_unsafe_realloc` is the root of the new functions. It is marked is
unsafe because direct usage is not recommended. The wrapper macros are
better suited for that.
It is possible to use `el_realloc` as a realloc replacement, the difference
is that `el_realloc` exits on errors and prints the file and line from
which is was called. Error reporting and exit behaviour can be customized
when using `el_custom_realloc`.