Improve unit tests for el_calloc functions
Remove `.exit_code = 0` in test cases. And exit code of zero is implicit
and doesn't have to be specified.

Addded three tests to `el_custom_calloc`, to verify that the error
function is called when it should, and isn't called when no error occurs.

Assigned and freed the pointer returned by `el_custom_calloc` when
`SIZE_MAX` is passed. This prevents a memory leak in no `exit()` is
called.

Added test to verify that a negative line number in `el_unsafe_calloc`
doesn't cause a crash.